Chat page is more developed.
I am understanding more about reusing variables and modules. I changed the layout.css file to base.css because it makes more sense to keep the most basic parts of my components together and the more dynamic parts of my page together. .page-container {} is now inside globals.css as it is being used for every page. I am a bit confused on one part. I have a component called screen container which is my whatsapp screen. My screen-container is a component of my page but i feel like i need to think about it more abstractly and consider it a template that I should move to globals.css despite it not being dynamic to all pages.